AND-878 move time-related unit tests to own package; add all method stubs;
