+++ src/test/java/com/onepagecrm/models/serializers/time/InstantSerializerTest.java
@@ -1,5 +1,6 @@
-package com.onepagecrm.models.serializers;
+package com.onepagecrm.models.serializers.time;
 
+import com.onepagecrm.models.serializers.InstantSerializer;
 import org.junit.Before;
 import org.junit.Test;
 import org.threeten.bp.Instant;
